model_training/pyproject.toml
[tool.poetry]
name = "feedback"
version = "0.1.0"
description = "Generation of feedback for novice helpers"
authors = ["Alicja"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
trl = "^0.15.2"
bitsandbytes = "^0.45.3"
peft = "^0.7.1"
scipy = "^1.11.4"
wandb = "^0.16.1"
torch = "^2.6.0"
python-dotenv = "^1.0.1"
transformers = "^4.49.0"
datasets = "^3.3.2"
torchvision = "^0.21.0"
numpy = "1.24"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
